pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "bettergit"
authors = [{name = "BetterGit Team", email = "team@bettergit.dev"}]
license = {file = "LICENSE"}
description = "A modern, intuitive version control system built on Git"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Version Control",
]
keywords = ["git", "version-control", "cli", "developer-tools"]
dependencies = [
    "click>=8.0.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "inquirer>=3.1.0",
    "keyring>=24.0.0",
    "requests>=2.28.0",
    "emoji>=2.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
bit = "bettergit.cli:main"

[project.urls]
Homepage = "https://github.com/bettergit/bettergit"
Documentation = "https://docs.bettergit.dev"
Repository = "https://github.com/bettergit/bettergit"
"Bug Tracker" = "https://github.com/bettergit/bettergit/issues"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=bettergit --cov-report=term-missing"
